refactor: move constants and tests
