+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const AddressZero = '0x0000000000000000000000000000000000000000'
