+++ src/index.ts
@@ -4,6 +4,3 @@ export * from './strategy/StrategyTree'
 export * from './strategy/utils'
 
 export * from './abi'
-
-export * from './types/index'
-export * from './types/helpers'
